+++ src/Razor/src/Microsoft.VisualStudio.LanguageServerClient.Razor/Endpoints/UpdateHtmlBuffer.cs
@@ -41,6 +41,6 @@ internal partial class RazorCustomMessageTarget
             hostDocumentUri,
             request.Changes.Select(change => change.ToVisualStudioTextChange()).ToArray(),
             request.HostDocumentVersion.Value,
-            state: null);
+            state: request.PreviousWasEmpty);
     }
 }
